Backend

+++ ZmitaCart.Backend/ZmitaCart.Application/Commands/OfferCommands/DeleteOffer/DeleteOfferHandler.cs
@@ -1,10 +1,12 @@
-﻿using MediatR;
+﻿using FluentResults;
+using MediatR;
+using ZmitaCart.Application.Common.Errors;
 using ZmitaCart.Application.Interfaces;
 using ZmitaCart.Application.Services;
 
 namespace ZmitaCart.Application.Commands.OfferCommands.DeleteOffer;
 
-public class DeleteOfferHandler : IRequestHandler<DeleteOfferCommand>
+public class DeleteOfferHandler : IRequestHandler<DeleteOfferCommand, Result>
 {
 	private readonly IOfferRepository _offerRepository;
 	private readonly ICurrentUserService _currentUserService;
@@ -17,17 +19,28 @@ public class DeleteOfferHandler : IRequestHandler<DeleteOfferCommand>
 		_pictureRepository = pictureRepository;
 	}
 
-	public async Task Handle(DeleteOfferCommand request, CancellationToken cancellationToken)
+	public async Task<Result> Handle(DeleteOfferCommand request, CancellationToken cancellationToken)
 	{
 		var user = _currentUserService.UserId;
 		if (user is null)
 		{
-			throw new UnauthorizedAccessException("You are not authorized to delete an offer.");
+			return Result.Fail(new UnauthorizedError("You are not authorized to delete an offer."));
 		}
 
 		var userId = int.Parse(user);
 		
-		await _pictureRepository.RemoveAsync(userId, request.Id);
-		await _offerRepository.DeleteAsync(userId, request.Id);
+		var removePic = await _pictureRepository.RemoveAsync(userId, request.Id);
+		if (removePic.IsFailed)
+		{
+			return Result.Fail(removePic.Errors);
+		}
+		
+		var removeOffer = await _offerRepository.DeleteAsync(userId, request.Id);
+		if (removeOffer.IsFailed)
+		{
+			return Result.Fail(removeOffer.Errors);
+		}
+
+		return Result.Ok();
 	}
 }
